+++ src/main/java/com/github/javafaker/CreditCardType.java
@@ -1,0 +1,15 @@
+package com.github.javafaker;
+
+public enum CreditCardType {
+    VISA,
+    MASTERCARD,
+    DISCOVER,
+    AMERICAN_EXPRESS,
+    DINERS_CLUB,
+    JCB,
+    SWITCH,
+    SOLO,
+    DANKORT,
+    FORBRUGSFORENINGEN,
+    LASER
+}